feat: Dynamic public path
This is mainly used in the tc-component, which is imported
into the showcase, which is served by a different server.

In order to keep the async module loading offered by webpack working,
we need to specify the hostname of the server serving the tc-component
and various split chunks (publicPath in webpack terms).

The future tc-component loading and injection will have to take that
into account.

Note: For some reason I don't understand, when configuring
optimization.splitChunks.chunks to 'all' instead of the default 'async'
(which is to say that all shared dependencies are to be split
in separate chunks to allow for more reuse), the eventBus throws
an error about duplicate callbacks; this might be because of some
reused code which is not supposed to.